pyproject.toml
[tool.ruff]
line-length = 120
target-version = "py311"

exclude =[
    "__pycache__",
    ".git",
    ".gitlab",
    ".gitlab-ci",
    "docker",
    "tests",
]

[tool.ruff.lint]
select = [
    "UP",   # pyupgrade
    "I",    # isort
    "E",    # flake8-errors
    "W",    # flake8-warnings
    "F",    # flake8-pyflakes
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "T10",  # flake8-debugger
    "PT",   # flake8-pytest-style
    "ANN",  # flake8-annotations
    "N",    # pep8-naming
    "RUF",  # Ruff
    "DTZ",  # flake8-datetimez
    "ERA",  # eradicate
    "FURB", # refurb
]

ignore = [
    "N812", # Lowercase imported as non-lowercase
    "ANN401",  # Any (allowed) - only when needed
]

[tool.ruff.format]
quote-style = "single"


[tool.poetry]
name = "style-transfer-streamlit"
version = "0.1.0"
description = "Style transfer app with Streamlit"
authors = ["phx7v"]
readme = "README.md"
packages = [{ include = "style_transfer" }]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.31"
torch = "^2.1"
pillow = "^10.2"
torchvision = "^0.16"
numpy = "^1.26"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
ruff = "^0.14"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
